added moment.js and created function
made a function that loops through all entries, and then adds the totals to an object depending on the entry's time stamp. no averages yet, but that should be easy. just need to add an html page for all this. might also add a new js document for functions. getting messy