WORKING
changed things so that there's an init function that sets global variables. Within this function we check localstorage, and set allcategories to whatever is in there, and if there's nothing , we set the preset value of two categories. Also changed html addition so that the categories don't get added in HTML until we hit new expense button.